renamed time to time_handling, don't shadow time
