+++ mslib/msui/mscolab_version_history.py
@@ -37,7 +37,7 @@ from PyQt5 import QtCore, QtWidgets, QtGui
 from mslib.msui.qt5 import ui_mscolab_version_history as ui
 from mslib.utils.qt import show_popup
 from mslib.utils.config import config_loader
-from mslib.utils.time import utc_to_local_datetime
+from mslib.utils.time_handling import utc_to_local_datetime
 
 
 class MSColabVersionHistory(QtWidgets.QMainWindow, ui.Ui_MscolabVersionHistory):
