+++ setup.cfg
@@ -39,4 +39,4 @@ norecursedirs = .git .idea .cache
 [flake8]
 ignore = E124,E125,E402,W504
 max-line-length = 120
-exclude = mslib/msui/qt5/*.py, mslib/mscolab/migrations/*.py
+exclude = mslib/msui/qt5/*.py, mslib/mscolab/migrations/*.py, mslib/plugins/*
